+++ gwcelery/conf/__init__.py
@@ -19,6 +19,11 @@ import os
 
 # Celery application settings.
 
+# Task tombstones expire after 5 minutes.
+# Celery's default setting of 1 day could cause the Redis database to grow too
+# large because we pass large byte strings as task arguments and return values.
+result_expires = 300
+
 # Use pickle serializer, because it supports byte values.
 accept_content = ['json', 'pickle']
 event_serializer = 'json'